Part I
You are required to add a functionality to the same list in the First Screen. The Second Screen now appears after a long press on any item.

The single click of an item will now only select/de-select the charity item from the list. (Please note, by default the first item should be selected. If user de-select all again the first item must be selected again.)

You need to add a swipe Left and swipe Right on the overall main activity(Screen) of the app.