Part II
Swipe Right takes you to a screen that has camera that can take picture only. The requirement for video/front-back/zoom will not be checked. The camera should only take a back picture and send button at the right bottom should open a pop-up. That’s it.

Swipe left from the Camera Screen back takes you to the main activity.